requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
pydantic>=2.12.0
python-dotenv>=1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart>=0.0.12
requests>=2.31.0
bcrypt==4.0.1
pg8000>=1.30.0
scramp>=1.4.4
supabase>=2.0.0